feat: volume meter 컴포넌트로 분리 #69
음성파형(볼륨 값 시각화)을 나타내는 컴포넌트를 별도로 분리했습니다.
