+++ frontend/src/pages/MicTest/MicTest.tsx
@@ -1,11 +1,9 @@
 import Header from "@/components/Header/Header";
-import AudioRecord from "./components/AudioRecord";
 
 const MicTest = () => {
   return (
     <>
       <Header type="instructor" />
-      <AudioRecord />
     </>
   );
 };
